requirements.txt
yfinance
requests
schedule
openai
pandas
python-dotenv
Flask